all cpp files commented

+++ splitscreen/src/save_manager.cpp
@@ -9,28 +9,36 @@
 #include <string>
 
 SaveManager::SaveManager() {
+  // open file
   std::fstream data("assets/save.txt");
   std::string line;
+  // read data into data variable line by line
   while (getline(data, line)) {
     data_.push_back(std::stoi(line));
   }
+  // close file
   data.close();
 }
 
 SaveManager::~SaveManager() {
+  // delete old save
   std::remove("assets/save.txt");
+  // create and open new save file
   std::ofstream data("assets/save.txt");
+  // input data into file from data_
   for (int i : data_) {
     data << std::to_string(i) << "\n";
   }
+  // clsoe file
   data.close();
 }
 
 unsigned int SaveManager::GetCredits() { return data_[0]; }
 
 unsigned int SaveManager::Reward() {
-  int avrgae_level = (data_[2] + data_[3] + data_[5] + data_[6]) / 4;
-  int reward = avrgae_level * 100;
+  // calculation of reward based on player level
+  float avrgae_level = (data_[2] + data_[3] + data_[5] + data_[6]) / 4.0f;
+  int reward = avrgae_level * 100.0f;
   data_[0] += reward;
   return reward;
 }
@@ -38,7 +46,9 @@ unsigned int SaveManager::Reward() {
 std::vector<int> SaveManager::GetGunStats(int type) {
   float level = data_[7 + type];
   float multiplier = level / 10;
+  // exponential cost based off level
   int cost = 10 * pow(1.1, level);
+  // linear incrase in upgrdae
   std::vector<int> stats = {
       static_cast<int>(static_cast<float>(kBaseGuns[type][0]) *
                        (1.0f + multiplier)),
@@ -49,6 +59,7 @@ std::vector<int> SaveManager::GetGunStats(int type) {
 bool SaveManager::UpgradeWeapon(int type) {
   int cost = GetGunStats(type)[3];
   bool bought = false;
+  // if suffuicent funds
   if (cost <= data_[0]) {
     data_[7 + type] += 1;
     data_[0] -= cost;
@@ -58,6 +69,7 @@ bool SaveManager::UpgradeWeapon(int type) {
 }
 
 std::vector<int> SaveManager::GetPlayerStats(int player) {
+  // setting index base off robot inputted
   int index = 2;
   if (player == 2) {
     index = 5;
@@ -74,6 +86,7 @@ std::vector<int> SaveManager::GetPlayerStats(int player) {
 }
 
 bool SaveManager::UpgradePlayer(bool upgrade, int player) {
+  // setting index base off robot inputted
   int index = 2;
   if (player == 2) {
     index = 5;
@@ -89,6 +102,7 @@ bool SaveManager::UpgradePlayer(bool upgrade, int player) {
 }
 
 int SaveManager::Repair(int player) {
+  // setting index base off robot inputted
   int index = 1;
   if (player == 2) {
     index = 4;
@@ -107,6 +121,7 @@ int SaveManager::Repair(int player) {
 }
 
 void SaveManager::SetMenuDamage(int player, int damage) {
+  // setting index base off robot inputted
   int index = 1;
   if (player == 2) {
     index = 4;
@@ -115,6 +130,7 @@ void SaveManager::SetMenuDamage(int player, int damage) {
 }
 
 int SaveManager::GetRepairCost(int player) {
+  // setting index base off robot inputted
   int index = 1;
   if (player == 2) {
     index = 4;
@@ -128,6 +144,7 @@ int SaveManager::GetRepairCost(int player) {
 }
 
 int SaveManager::GetDamage(int player) {
+  // setting index base off robot inputted
   int index = 1;
   if (player == 2) {
     index = 4;
@@ -136,5 +153,6 @@ int SaveManager::GetDamage(int player) {
 }
 
 int SaveManager::GetLevel() {
+  // calculates the mean level of the robots
   return (data_[2] + data_[3] + data_[5] + data_[6]) / 4;
 }
